backend/core/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
SQLAlchemy
asyncpg
python-jose[cryptography]
httpx